requirements.txt
openai
python-dotenv
numpy<2.0
faiss-cpu
pandas


langchain
langchain-openai
langchain-community
pytest